+++ examples/pong_tutorial_04/pong.rs
@@ -21,7 +21,7 @@ pub const BALL_RADIUS: f32 = 2.0;
 
 pub struct Pong;
 
-impl<'a, 'b> SimpleState<'a, 'b> for Pong {
+impl SimpleState for Pong {
     fn on_start(&mut self, data: StateData<GameData>) {
         let world = data.world;